Version 0.4.4
implement list of active/dead sensors to avoid continuous logging of
dead sensors,
also added some ´fool proof´ error handlers based on users feedback

+++ plugin.py
@@ -1,11 +1,9 @@
 """
 Smart Virtual Thermostat python plugin for Domoticz
-
 Author: Logread,
         adapted from the Vera plugin by Antor, see:
             http://www.antor.fr/apps/smart-virtual-thermostat-eng-2/?lang=en
             https://github.com/AntorFr/SmartVT
-
 Version:    0.0.1: alpha
             0.0.2: beta, with new connection object from Domoticz Python plugin framework
             0.0.3: bug fixes + added thermostat temp update even if in Off or Forced modes
@@ -51,10 +49,11 @@ Version:    0.0.1: alpha
             0.4.3: Call domoticz json API at least every 5 minutes to avoid the 10 mins timout of connections that
                     floods domoticz's log with "Incoming Connection from ...." messages. Take advantage of this to
                     update the temperature readings more often than upon each calculation
-
+            0.4.4: implement list of active/dead sensors to avoid continuous logging of dead sensors,
+            		also added some ´fool proof´ error handlers based on users feedback
 """
 """
-<plugin key="SVT" name="Smart Virtual Thermostat" author="logread" version="0.4.3" wikilink="https://www.domoticz.com/wiki/Plugins/Smart_Virtual_Thermostat.html" externallink="https://github.com/999LV/SmartVirtualThermostat.git">
+<plugin key="SVT" name="Smart Virtual Thermostat" author="logread" version="0.4.4" wikilink="https://www.domoticz.com/wiki/Plugins/Smart_Virtual_Thermostat.html" externallink="https://github.com/999LV/SmartVirtualThermostat.git">
     <description>
         <h2>Smart Virtual Thermostat</h2><br/>
         Easily implement in Domoticz an advanced virtual thermostat based on time modulation<br/>
@@ -101,6 +100,7 @@ import urllib.request as request
 from datetime import datetime, timedelta
 import time
 import base64
+import itertools
 
 class deviceparam:
 
@@ -121,6 +121,7 @@ class BasePlugin:
         self.pauseondelay = 2  # time between pause sensor actuation and actual pause
         self.pauseoffdelay = 1  # time between end of pause sensor actuation and end of actual pause
         self.forcedduration = 60  # time in minutes for the forced mode
+        self.ActiveSensors = {}
         self.InTempSensors = []
         self.OutTempSensors = []
         self.Heaters = []
@@ -211,6 +212,10 @@ class BasePlugin:
         Domoticz.Debug("Outside Temperature sensors = {}".format(self.OutTempSensors))
         self.Heaters = parseCSV(Parameters["Mode3"])
         Domoticz.Debug("Heaters = {}".format(self.Heaters))
+        
+        # build dict of status of all temp sensors to be used when handling timeouts
+        for sensor in itertools.chain(self.InTempSensors, self.OutTempSensors):
+            self.ActiveSensors[sensor] = True
 
         # splits additional parameters
         params = parseCSV(Parameters["Mode5"])
@@ -275,6 +280,11 @@ class BasePlugin:
     def onHeartbeat(self):
 
         now = datetime.now()
+        
+        # fool proof checking.... based on users feedback
+        if not all(device in Devices for device in (1,2,3,4,5,6)):
+            Domoticz.Error("one or more devices required by the plugin is/are missing, please check domoticz device creation settings and restart !")
+            return
 
         if Devices[1].sValue == "0":  # Thermostat is off
             if self.forced or self.heat:  # thermostat setting was just changed so we kill the heating
@@ -458,6 +468,11 @@ class BasePlugin:
                     else:
                         Domoticz.Error("Device with idx={} does not seem to be a switch !".format(idx))
 
+        # fool proof checking.... based on users feedback
+        if len(switches) == 0:
+            Domoticz.Error("none of the devices in the 'heaters' parameter is a switch... no action !")
+            return
+
         # flip on / off as needed
         self.heat = switch
         command = "On" if switch else "Off"
@@ -486,20 +501,16 @@ class BasePlugin:
                     if "Temp" in device:
                         Domoticz.Debug("device: {}-{} = {}".format(device["idx"], device["Name"], device["Temp"]))
                         # check temp sensor is not timed out
-                        if not SensorTimedOut(device["LastUpdate"]):
+                        if not self.SensorTimedOut(idx, device["Name"], device["LastUpdate"]):
                             listintemps.append(device["Temp"])
-                        else:
-                            Domoticz.Error("skipping timed out temperature sensor {}".format(device["Name"]))
                     else:
                         Domoticz.Error("device: {}-{} is not a Temperature sensor".format(device["idx"], device["Name"]))
                 elif idx in self.OutTempSensors:
                     if "Temp" in device:
                         Domoticz.Debug("device: {}-{} = {}".format(device["idx"], device["Name"], device["Temp"]))
                         # check temp sensor is not timed out
-                        if not SensorTimedOut(device["LastUpdate"]):
+                        if not self.SensorTimedOut(idx, device["Name"], device["LastUpdate"]):
                             listouttemps.append(device["Temp"])
-                        else:
-                            Domoticz.Error("skipping timed out temperature sensor {}".format(device["Name"]))
                     else:
                         Domoticz.Error("device: {}-{} is not a Temperature sensor".format(device["idx"], device["Name"]))
 
@@ -571,6 +582,31 @@ class BasePlugin:
         elif level == "Normal":
             Domoticz.Log(message)
 
+    def SensorTimedOut(self, idx, name, datestring):
+
+        def LastUpdate(datestring):
+            dateformat = "%Y-%m-%d %H:%M:%S"
+            # the below try/except is meant to address an intermittent python bug in some embedded systems
+            try:
+                result = datetime.strptime(datestring, dateformat)
+            except TypeError:
+                result = datetime(*(time.strptime(datestring, dateformat)[0:6]))
+            return result
+
+        timedout = LastUpdate(datestring) + timedelta(minutes=int(Settings["SensorTimeout"])) < datetime.now()
+
+        # handle logging of time outs... only log when status changes (less clutter in logs)
+        if timedout:
+            if self.ActiveSensors[idx]:
+                Domoticz.Error("skipping timed out temperature sensor '{}'".format(name))
+                self.ActiveSensors[idx] = False
+        else:
+            if not self.ActiveSensors[idx]:
+                Domoticz.Status("previously timed out temperature sensor '{}' is back online".format(name))
+                self.ActiveSensors[idx] = True
+
+        return timedout
+
 
 global _plugin
 _plugin = BasePlugin()
@@ -597,21 +633,6 @@ def onHeartbeat():
 
 
 # Plugin utility functions ---------------------------------------------------
-
-
-def SensorTimedOut(datestring):
-
-    def LastUpdate(datestring):
-        dateformat = "%Y-%m-%d %H:%M:%S"
-        # the below try/except is meant to address an intermittent python bug in some embedded systems
-        try:
-            result = datetime.strptime(datestring, dateformat)
-        except TypeError:
-            result = datetime(*(time.strptime(datestring, dateformat)[0:6]))
-        return result
-
-    return LastUpdate(datestring) + timedelta(minutes=int(Settings["SensorTimeout"])) < datetime.now()
-
 
 def parseCSV(strCSV):
